Add login (fixes #2)
An unauthenticated user will be redirected to a login page. Entering
correct credentials will redirect them to the home page with an auth
token set for future authenticated calls. A non-200 from the back end
will display an error message.

This slightly refactors the router/views in ways I would be open to
changing.