Enable hot reload
This changes `make run-local` to install any dependencies for both the
app and api and starts both with hot reloading. The go server played
nicer with sending to the background so I started it first. This also
adds some minimal documentation in the README.

This almost certainly won't work on Windows, but should on Mac/Linux (I
tested on LMDE 3 Cindy). Windows development is something we'll need to
support at some point in the interest of being accessible to anyone who
wants to help.

Fixes #32