Add internationalization
I went with vue-i18n, as it seemed like the best option. I'm fairly sure
I got everything, but we should probably move everything out of views
and into components (like with Home).

I have no idea how we're going to get translations for non-English
languages: we'll have to rely on PRs for that.

Closes #42