Add /logout
As there's no session table on the back end this is a vue-only change.
It also changes /login to skip the flow if an auth token is present.
Dealing with auth tokens that aren't valid JWTs is handled in #34.

Currently logout will redirect a user to /, which redirects to the login
page. That should change, but as it stands the logout logic won't need
to when it does.

Closes #33