read timestamps/read anywhere (#120)
The "I've read this" button on the book detail view wouldn't actually do
anything, as it emitted an event that wasn't listened to. This moves the
logic of reading a book to the `Book` component, which displays an error
if the axios object hasn't been provided. It's not needed in the list
view, for instance, because you currently can't read/review from there.