More federation work (#6)
This is 80% or so of the way there to federate reads. The issue is that
you more or less have to get it completely working to get it partially
working. I decided to stop when I realize I would be changing how we
create reads: I think we should save the actual insertion of the Read
objects into the database for in the AP flow. This way there is only one
way for a Read to be entered, regardless of it's local or federated.

I've left in the debugging log lines for now to help understand what's
happening where. Before 6 is closed I'd like for those to be removed.